requirements.txt
azure-identity
azure-storage-blob
cosmotech-api
kubernetes
